JiHeeChoi22.github.io/Automation/appium_test_project/requirements.txt
Appium-Python-Client>=2.0.0
pytest>=7.0.0
pytest-html>=3.1.1
selenium>=4.0.0
PyYAML>=6.0.0
python-json-logger>=2.0.7
pytest-xdist>=3.0.0